refactor: storage独立のコンポーネントにした
