+++ Frontend/src/components/date-registration/RegistrationList.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Text, View } from "react-native";
 import { useTailwind } from "tailwind-rn";
-import { storage } from "src/components/date-registration/DateStorage";
+import { storage } from "src/components/storage/Storage";
 
 const RegistrationList = () => {
   const tailwind = useTailwind();
